+++ yash-syntax/CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Items in the `source` module:
     - `Line` renamed to `Code`
     - `Location`'s field `line` renamed to `code`
+    - Type of the following fields from `Location` to `LocationRef`:
+        - `Param::location`
+    - Parameter and return type of `WordLexer::braced_param` from `Location` to `LocationRef`
 - Dependency versions
     - `async-trait` 0.1.50 → 0.1.52
     - `futures-util` 0.3.18 → 0.3.19